mechanical move of the iterator to its own file
